requirements.txt
falcon==0.3.0
GitPython==1.0.1
nose==1.3.7		# for (py-)pgp tests

# Using both (py)gpgme and python-pgp because the former is faster and
# more secure, but the second allows me more fine grained control.
-e git://github.com/wdoekes/python-pgp@wjd#egg=pgp
pygpgme==0.3		# python3-gpgme (req: libgpgme11-dev)

# PGP:
# - not OpenPGP (http://www.aonalu.net/openpgp)
# - not ??? (https://github.com/singpolyma/OpenPGP-Python)
# - not mitchellrj-python-pgp because of wjd pull requests
#   (git://github.com/mitchellrj/python-pgp#egg=pgp)
# - not py-pgp (a fork of mitchellrj-python-pgp for the same
#   reason)
# - optional dep: camcrypt (for Camellia)
# - optional dep: twofish (for Twofish) -- apt-get install libtwofish-dev

# Dependencies of python-pgp v@wjd:
#pycrypto==2.6.1
#python-magic==0.4.10
#requests==2.9.1
#six==1.10.0

# Dependencies of GitPython v1.0.1:
#gitdb==0.6.4
#smmap==0.9.0

# Dependencies of falcon v0.3.0:
#python-mimeparse==0.1.4
